requirements.txt
pandas
google-cloud-bigquery
google-cloud-storage
python-dotenv
pyarrow  # Required for handling Parquet files
openpyxl  # Required for handling Excel files
